release 0.1.83 source code

+++ services/eihealth/v1/model/model_bounding_box_dto.go
@@ -16,7 +16,7 @@ type BoundingBoxDto struct {
 	Size []float32 `json:"size"`
 
 	// 填充
-	Padding *int32 `json:"padding,omitempty"`
+	Padding *float32 `json:"padding,omitempty"`
 }
 
 func (o BoundingBoxDto) String() string {
